+++ xfields/__init__.py
@@ -24,8 +24,8 @@ from .beam_elements.beambeam3d import ConfigForUpdateBeamBeamBiGaussian3D
 from .beam_elements.temp_slicer import TempSlicer
 from .beam_elements.electroncloud import ElectronCloud
 from .beam_elements.electronlens_interpolated import ElectronLensInterpolated
-from .beam_elements.wakefield import WakeComponent, ResonatorWake
-from .beam_elements.wakefield import Wakefield
+from .beam_elements.wakefields import WakeComponent, ResonatorWake
+from .beam_elements.wakefields import Wakefield
 from .beam_elements.transverse_damper import TransverseDamper
 
 from .general import _pkg_root
