+++ xfields/beam_elements/wakefields/__init__.py
@@ -1,0 +1,2 @@
+from .wakefield import WakeComponent, ResonatorWake
+from .wakefield import Wakefield
